+++ lib/kitchen/directions/bake_eoc_section_content/use_section_title.rb
@@ -1,0 +1,25 @@
+# frozen_string_literal: true
+
+module Kitchen
+  module Directions
+    module UseSectionTitle
+      def self.v1(chapter:, eoc_selector: '', section_selector: '')
+        composite_page = chapter.search(eoc_selector).first
+        section = composite_page.search(section_selector).first
+
+        section_title = section.first('[data-type="title"]')
+
+        composite_page_title = composite_page.first('[data-type="document-title"]')
+
+        composite_page_title.append(child:
+          <<~HTML
+            <span class="os-divider"> - </span>
+            <span class="os-subtitle">#{section_title.text}</span>
+          HTML
+        )
+
+        section_title.trash
+      end
+    end
+  end
+end