+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* Create `UseSectionTitle` to move section title to composite page header
 * Remove `ElementBase#copied_id` since we don't copy IDs anymore
 
 ## [v1.22.0] - 2023-01-13